auth_service/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.5
pydantic==2.8.2
PyJWT==2.9.0
passlib==1.7.4
email-validator==2.2.0
python-dotenv==1.0.1
